% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_info_combo2.R
\docType{data}
\name{dose_info_combo2}
\alias{dose_info_combo2}
\title{Dataset: trial dose information for a dual-agent combination study}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 33 rows and 4 columns.
}
\usage{
dose_info_combo2
}
\description{
The data set defines all pre-defined dose-levels which can be
explored in the dual-agent example trial.
}
\details{
\describe{

\item{group_id}{study}

\item{drug_A}{drug A dose amount}

\item{drug_B}{drug B dose amount}

\item{dose_id}{unique id of record}

}
}
\seealso{
Other blrm_trial combo2 example: 
\code{\link{blrm_trial}()},
\code{\link{drug_info_combo2}},
\code{\link{example-combo2_trial}}
}
\concept{blrm_trial combo2 example}
\keyword{datasets}
