% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_data.R
\name{.get_strata_group_fct}
\alias{.get_strata_group_fct}
\title{extracts from a blrmfit object and a given data-set the group and
stratum factor}
\usage{
.get_strata_group_fct(object, data)
}
\description{
extracts from a blrmfit object and a given data-set the group and
stratum factor
}
\keyword{internal}
