% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blrm_exnex.R
\name{.label_index}
\alias{.label_index}
\title{Utility function to label parameter indices according to factor
levels.}
\usage{
.label_index(stanfit, par, ...)
}
\arguments{
\item{stanfit}{stan fit which names are being modified}

\item{par}{parameter selected}

\item{...}{must include as many factors as there are indices which
are used in the order given to translate indices to text labels}
}
\description{
Utility function to label parameter indices according to factor
levels.
}
\keyword{internal}
