% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blrm_exnex.R
\name{.validate_group_stratum_nesting}
\alias{.validate_group_stratum_nesting}
\title{Test if each group is assigned to exactly 1 stratum. Error
otherwise.}
\usage{
.validate_group_stratum_nesting(group_def, strata_def)
}
\arguments{
\item{group_def}{groups of data}

\item{strata_def}{assigned stratum to group}
}
\description{
Test if each group is assigned to exactly 1 stratum. Error
otherwise.
}
\keyword{internal}
