% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{draws-OncoBayes2}
\alias{draws-OncoBayes2}
\alias{as_draws}
\alias{as_draws_matrix}
\alias{as_draws_array}
\alias{as_draws_df}
\alias{as_draws_rvars}
\alias{as_draws_list}
\alias{as_draws.blrmfit}
\alias{as_draws_matrix.blrmfit}
\alias{as_draws_array.blrmfit}
\alias{as_draws_df.blrmfit}
\alias{as_draws_list.blrmfit}
\alias{as_draws_rvars.blrmfit}
\alias{as_draws.blrm_trial}
\alias{as_draws_matrix.blrm_trial}
\alias{as_draws_array.blrm_trial}
\alias{as_draws_df.blrm_trial}
\alias{as_draws_list.blrm_trial}
\alias{as_draws_rvars.blrm_trial}
\title{Transform \code{blrmfit} or \code{blrm_trial} to \code{draws} objects}
\usage{
\method{as_draws}{blrmfit}(x, variable = NULL, regex = FALSE, inc_warmup = FALSE, ...)

\method{as_draws_matrix}{blrmfit}(x, variable = NULL, regex = FALSE, inc_warmup = FALSE, ...)

\method{as_draws_array}{blrmfit}(x, variable = NULL, regex = FALSE, inc_warmup = FALSE, ...)

\method{as_draws_df}{blrmfit}(x, variable = NULL, regex = FALSE, inc_warmup = FALSE, ...)

\method{as_draws_list}{blrmfit}(x, variable = NULL, regex = FALSE, inc_warmup = FALSE, ...)

\method{as_draws_rvars}{blrmfit}(x, variable = NULL, regex = FALSE, inc_warmup = FALSE, ...)

\method{as_draws}{blrm_trial}(x, variable = NULL, regex = FALSE, inc_warmup = FALSE, ...)

\method{as_draws_matrix}{blrm_trial}(x, variable = NULL, regex = FALSE, inc_warmup = FALSE, ...)

\method{as_draws_array}{blrm_trial}(x, variable = NULL, regex = FALSE, inc_warmup = FALSE, ...)

\method{as_draws_df}{blrm_trial}(x, variable = NULL, regex = FALSE, inc_warmup = FALSE, ...)

\method{as_draws_list}{blrm_trial}(x, variable = NULL, regex = FALSE, inc_warmup = FALSE, ...)

\method{as_draws_rvars}{blrm_trial}(x, variable = NULL, regex = FALSE, inc_warmup = FALSE, ...)
}
\arguments{
\item{x}{A \code{blrmfit} or \code{blrm_trial} object.}

\item{variable}{A character vector providing the variables to
extract.  By default, all variables are extracted.}

\item{regex}{Logical; Should variable be treated as a (vector of)
regular expressions? Any variable in \code{x} matching at least
one of the regular expressions will be selected. Defaults to
\code{FALSE}.}

\item{inc_warmup}{Should warmup draws be included? Defaults to
\code{FALSE}.}

\item{...}{Arguments passed to individual methods (if applicable).}
}
\description{
Transform a \code{blrmfit} or \code{blrm_trial} object to a format supported by the
\pkg{posterior} package.
}
\details{
To subset iterations, chains, or draws, use the
\code{\link[posterior:subset_draws]{posterior::subset_draws()}} method after
transforming the input object to a \code{draws} object.
}
\examples{
## Setting up dummy sampling for fast execution of example
## Please use 4 chains and 100x more warmup & iter in practice
.user_mc_options <- options(
  OncoBayes2.MC.warmup = 10, OncoBayes2.MC.iter = 20, OncoBayes2.MC.chains = 1,
  OncoBayes2.MC.save_warmup = FALSE
)

# fit an example model. See documentation for "combo2" example
example_model("combo2")

post <- as_draws(blrmfit)

## Recover user set sampling defaults
options(.user_mc_options)

}
\seealso{
\code{\link[posterior:draws]{posterior::draws()}}
\code{\link[posterior:subset_draws]{posterior::subset_draws()}}
}
