% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_info_combo2.R
\docType{data}
\name{drug_info_combo2}
\alias{drug_info_combo2}
\title{Dataset: drug information for a dual-agent combination study}
\format{
A tibble with 2 rows (one per durg) and 4 columns:

\describe{
  \item{drug_name}{name of drug}
  \item{dose_ref}{reference dose}
  \item{dose_unit}{units used for drug amounts}
  \item{reference_p_dlt}{a-priori probability for a DLT at the reference dose}
}
}
\usage{
drug_info_combo2
}
\description{
Data set describing the two drugs involved in the example for a
dual-agent combination study.
}
\seealso{
Other blrm_trial combo2 example: 
\code{\link{blrm_trial}()},
\code{\link{dose_info_combo2}},
\code{\link{example-combo2_trial}}
}
\concept{blrm_trial combo2 example}
\keyword{datasets}
