% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsamples.R
\name{nsamples.blrmfit}
\alias{nsamples.blrmfit}
\alias{nsamples}
\title{Return the number of posterior samples}
\usage{
\method{nsamples}{blrmfit}(object, ...)
}
\arguments{
\item{object}{fitted model object}

\item{...}{not used in this function}
}
\description{
Return the number of posterior samples
}
\examples{
## Setting up dummy sampling for fast execution of example
## Please use 4 chains and 100x more warmup & iter in practice
.user_mc_options <- options(OncoBayes2.MC.warmup=10, OncoBayes2.MC.iter=20, OncoBayes2.MC.chains=1)


## run single-agent analysis which defines blrmfit model object
example_model("single_agent")

nsamples(blrmfit)

## Recover user set sampling defaults
options(.user_mc_options)

}
