% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_data.R
\name{pp_data}
\alias{pp_data}
\title{Internal function to simulate from the posterior new parameter draws}
\usage{
pp_data(object, newdata, draws, re.form)
}
\description{
Internal function to simulate from the posterior new parameter draws
}
\keyword{internal}
